Fix go validator
github.com/linode/linodego yanked 1.0.0, so this runs a go mod tidy and bumps
it to 1.1.0

We also do a bit more tracking here, introducing a couple of new values to the validator
struct to keep track of the labelsets within the current metric family. For each new sample,
we see whether its labelset (minus the expected labels for the given metric name, e.g. "le" for a histogram _bucket)
is the same as the previous sample. If _not_ then we check whether we've seen the labelset before and
bail if so as the family is interweaved