Fix bad_grouping_or_ordering_0 test (#232)
* Fix bad_grouping_or_ordering_0 test

This test is really janky in that it actually tests 4 things,
rather than just the metric interweaving it proports to test:

- "A Histogram MetricPoint MUST contain at least one bucket (the given one doesn't)
- "Histogram MetricPoints MUST have at least a bucket with an +Inf threshold" (the given one doesn't)
- The sum without a count thing I mentioned in #229 (the histograms have sums without counts)
- The metric interweaving (what it's mean't to be testing)

This commit fixes the first three to make this only test one thing

* Fix go validator

github.com/linode/linodego yanked 1.0.0, so this runs a go mod tidy and bumps
it to 1.1.0

We also do a bit more tracking here, introducing a couple of new values to the validator
struct to keep track of the labelsets within the current metric family. For each new sample,
we see whether its labelset (minus the expected labels for the given metric name, e.g. "le" for a histogram _bucket)
is the same as the previous sample. If _not_ then we check whether we've seen the labelset before and
bail if so as the family is interweaved