protobuf: Improve naming of `oneof` fields
The name after the `oneof` doesn't show up in the explicit getter for
a particular field, e.g. for the sum of a `HistogramValue` as a
double, the getter is called `GetDoubleValue` in Go. Nothing in this
getter tells you that you are getting the sum. For one, this is
confusing. But you will also get a name collision if any of the other
fields becomes a `oneof` (e.g. the count for a float histogram). With
this commit, the getter will be called `GetDoubleSum` etc.

Signed-off-by: beorn7 <beorn@grafana.com>