protobuf: Add "float histogram" capability
There are rare use cases of "weighted" or "scaled" histogram, where
the total count and the count in buckets are actually floating point
values.

The need for this "float histogram" becomes more obvious with the
introduction of Native Histograms. With Histograms being first class
citizens also within Prometheus, a recording rule that outputs
Histograms will generally output Histograms where the counts are
floating point numbers again (because PromQL operates on floats only,
and because the result of a `rate` operation is interpolated and
therefore not integral anymore, even if the input is). If such a
result of a recording rule is federated, we need to represent "float
histograms" in the exposition format.

This commit introduces the "float histogram" capability to the
conventional Histogram, which is also needed to add Native Histograms
later.

Signed-off-by: beorn7 <beorn@grafana.com>